+++ src/auth0__env__validate/index.js
@@ -1,7 +1,7 @@
-import { missing_env__throw } from '@ctx-core/env'
+import { import_meta_env_, missing_env__throw } from '@ctx-core/env'
 export function auth0__env__validate() {
-	if (!process.env.AUTH0_CLIENT_ID) missing_env__throw('AUTH0_CLIENT_ID')
-	if (!process.env.AUTH0_DOMAIN) missing_env__throw('AUTH0_DOMAIN')
+	if (!import_meta_env_().AUTH0_CLIENT_ID) missing_env__throw('AUTH0_CLIENT_ID')
+	if (!import_meta_env_().AUTH0_DOMAIN) missing_env__throw('AUTH0_DOMAIN')
 }
 export {
 	auth0__env__validate as validate_auth0_env,