+++ src/auth0__init/index.js
@@ -1,6 +1,7 @@
 ///public
 import { AUTH0_CLIENT_ID__set } from '../AUTH0_CLIENT_ID__/index.js'
 import { AUTH0_DOMAIN__set } from '../AUTH0_DOMAIN__/index.js'
+import { import_meta_env_ } from '@ctx-core/env'
 /**
  * @param ctx{import('@ctx-core/object').Ctx}
  * @param props{import('./index.d.ts').auth0__init__props_T}
@@ -8,6 +9,6 @@ import { AUTH0_DOMAIN__set } from '../AUTH0_DOMAIN__/index.js'
 export function auth0__init(
 	ctx, props = {}
 ) {
-	AUTH0_CLIENT_ID__set(ctx, props.AUTH0_CLIENT_ID || process.env.AUTH0_CLIENT_ID)
-	AUTH0_DOMAIN__set(ctx, props.AUTH0_DOMAIN || process.env.AUTH0_DOMAIN)
+	AUTH0_CLIENT_ID__set(ctx, props.AUTH0_CLIENT_ID || import_meta_env_().AUTH0_CLIENT_ID)
+	AUTH0_DOMAIN__set(ctx, props.AUTH0_DOMAIN || import_meta_env_().AUTH0_DOMAIN)
 }