+++ src/auth0__passwordless_start__fetch_post/index.d.ts
@@ -1,20 +1,34 @@
 import type { Ctx } from '@ctx-core/object'
 import type { Auth0Error } from 'auth0-js'
-import type { auth0__client_id__optional_body_T } from '../_types'
 export declare function auth0__passwordless_start__fetch_post(
-	ctx:Ctx, body:auth0__passwordless_start__fetch__body_T
+	ctx:Ctx, body:auth0__passwordless_start__fetch_post__body_T
 ):Promise<[null|Auth0Error, Response]>
 export {
 	auth0__passwordless_start__fetch_post as post_auth0_passwordless_start,
 }
-export interface auth0__passwordless_start__fetch__optional_body_T extends auth0__client_id__optional_body_T {
-	connection:string
-	send:string
-	email:string
-}
-export declare type post_auth0_passwordless_start_optional_body_T = auth0__passwordless_start__fetch__optional_body_T
-export interface auth0__passwordless_start__fetch__body_T extends auth0__passwordless_start__fetch__optional_body_T {
+//@formatter:off
+/**
+ * @see https://auth0.com/docs/authenticate/passwordless/implement-login/embedded-login/relevant-api-endpoints#post-passwordless-start
+ */
+export type auth0__passwordless_start__fetch_post__body_T = {
 	client_id:string
-}
-export declare type post_auth0_passwordless_start_body_T = auth0__passwordless_start__fetch__body_T
+	client_secret?:string
+	connection:'email'|'sms'
+	send:'link'|'code'
+	authParams?:{
+		scope?:'openid'
+		state?:string
+		redirect_uri?:string
+	}
+}&({
+		connection:'email'
+		email:string
+	}
+	|{
+		connection:'sms'
+		phone_number:string
+	})
+//@formatter:on
+export declare type auth0__passwordless_start__fetch__body_T = auth0__passwordless_start__fetch_post__body_T
+export declare type post_auth0_passwordless_start_body_T = auth0__passwordless_start__fetch_post__body_T
 export declare type post_auth0_passwordless_start_T = typeof auth0__passwordless_start__fetch_post